Screensaver: Did some stuff
I had been trying to fix the slow-down problem, but I don't think I can do that without either fractional delta values or change the code to use vectors vs. simple delta values.

What remains is a random number of savers by default, a potential performance boost in using `fillRect()`, and a placeholder to add and remove to the number of savers with the keyboard.